Table resizing extracted to a separate file
